move osquery packages back under /osquery
