+++ pkg/osquery/runtime/runtime.go
@@ -1,4 +1,4 @@
-package osqruntime
+package runtime
 
 import (
 	"context"
@@ -18,7 +18,7 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/kolide/launcher/pkg/osqtable"
+	"github.com/kolide/launcher/pkg/osquery/table"
 	"github.com/kolide/osquery-go"
 	"github.com/kolide/osquery-go/plugin/config"
 	"github.com/kolide/osquery-go/plugin/distributed"
@@ -533,7 +533,7 @@ func (r *Runner) launchOsqueryInstance() error {
 	}
 
 	plugins := o.opts.extensionPlugins
-	for _, t := range osqtable.PlatformTables(o.extensionManagerClient, o.logger) {
+	for _, t := range table.PlatformTables(o.extensionManagerClient, o.logger) {
 		plugins = append(plugins, t)
 	}
 	o.extensionManagerServer.RegisterPlugin(plugins...)
