+++ pkg/osquery/table/SUSharedPrefs.h
