+++ pkg/osquery/table/airdrop_darwin.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
