+++ pkg/osquery/table/best_practices.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
