+++ pkg/osquery/table/best_practices_test.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
