+++ pkg/osquery/table/chrome_state_darwin.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"io/ioutil"
@@ -7,8 +7,8 @@ import (
 )
 
 const (
-	usersDir = `C:\Users`
-	userData = `AppData\Local\Google\Chrome\User Data\Local State`
+	usersDir = "/Users"
+	userData = "Library/Application Support/Google/Chrome/Local State"
 )
 
 func findChromeStateFiles() []string {
@@ -17,9 +17,8 @@ func findChromeStateFiles() []string {
 	if err != nil {
 		return []string{}
 	}
-
 	for _, f := range filesInUser {
-		if f.IsDir() && (f.Name() != "Public") {
+		if f.IsDir() && (f.Name() != "Guest" || f.Name() != "Shared") {
 			stateFilePath := filepath.Join(usersDir, f.Name(), userData)
 			if _, err := os.Stat(stateFilePath); err == nil {
 				chromeLocalStateFiles = append(chromeLocalStateFiles, stateFilePath)
