+++ pkg/osquery/table/chrome_state_linux.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"io/ioutil"
