+++ pkg/osquery/table/email_addresses.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
