+++ pkg/osquery/table/kolide_vulnerabilities.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
