+++ pkg/osquery/table/launcher_identifier.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
