+++ pkg/osquery/table/macho.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"context"
