+++ pkg/osquery/table/spotlight.go
@@ -1,6 +1,6 @@
 // +build darwin
 
-package osqtable
+package table
 
 import (
 	"context"
