+++ pkg/osquery/table/sus_darwin.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 /*
 #cgo darwin CFLAGS: -DDARWIN -x objective-c
