+++ pkg/osquery/table/table.go
@@ -1,4 +1,4 @@
-package osqtable
+package table
 
 import (
 	"github.com/go-kit/kit/log"
